% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.zoningout.R
\docType{package}
\name{expfactory.zoningout}
\alias{expfactory.zoningout}
\title{expfactory.zoningout: A package for processing zoning out task data.}
\description{
This package processes data generated from a reading task for use in mind-wandering
experiments. The package can analyse data the task which runs in a web browser, and
was designed to be used within The Experiment Factory framework.
}
\section{expfactory.zoningout functions}{

...
}

\references{
Sayette, M. A., Reichle, E. D., & Schooler, J. W. (2009). Lost in the sauce: The effects of alcohol on mind wandering.
Psychological Science, 20(6), 747–752.
}
