% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.zoningout.R
\name{process_zoning_out_files}
\alias{process_zoning_out_files}
\title{Process Experiment Factory zoning out data files}
\usage{
process_zoning_out_files(path)
}
\arguments{
\item{path}{Path to data files}
}
\value{
Data frame
}
\description{
Process Experiment Factory zoning data files.
}
